UIORGS-285-add-type-to-edit-create-view outsource FieldOrganizationTypes
