+++ src/Organizations/OrganizationForm/OrganizationSummaryForm/FieldOrganizationTypes/index.js
@@ -1,0 +1,1 @@
+export * from './FieldOrganizationTypes';
