+++ src/Organizations/OrganizationForm/OrganizationSummaryForm/OrganizationSummaryForm.js
@@ -1,10 +1,6 @@
 import React, { useCallback, useState } from 'react';
 import { FormattedMessage } from 'react-intl';
 import PropTypes from 'prop-types';
-import {
-  find,
-  map,
-} from 'lodash';
 import { Field, useForm } from 'react-final-form';
 import { FieldArray } from 'react-final-form-arrays';
 
@@ -20,12 +16,12 @@ import {
 } from '@folio/stripes/components';
 import {
   AcqUnitsField,
-  FieldMultiSelectionFinal,
   validateRequired,
 } from '@folio/stripes-acq-components';
 
 import { ORGANIZATION_STATUS } from '../../../common/constants';
 import { FieldLanguage } from '../../../common/components';
+import { FieldOrganizationTypes } from './FieldOrganizationTypes';
 import {
   CREATE_UNITS_PERM,
   MANAGE_UNITS_PERM,
@@ -89,31 +85,7 @@ function OrganizationSummaryForm({ initialValues, organizationTypes, mutators })
 
   const isEditMode = Boolean(initialValues.id);
 
-  const getSelectedTypes = (e) => {
-    mutators.setType({}, e);
-  };
-
-  const formatter = ({ option }) => {
-    const item = find(organizationTypes, { id: option }) || option;
-
-    if (!item) return option;
-
-    return item.name;
-  };
-
-  const itemToString = item => item;
-
-  const typeOptions = map(organizationTypes, 'id');
-
-  const filter = (filterText, list) => {
-    const renderedItems = filterText
-      ? organizationTypes
-        .filter(group => group.name.toLowerCase().includes(filterText.toLowerCase()))
-        .map(({ id }) => id)
-      : list;
-
-    return { renderedItems };
-  };
+  const _organizationTypes = organizationTypes;
 
   return (
     <Row>
@@ -178,16 +150,9 @@ function OrganizationSummaryForm({ initialValues, organizationTypes, mutators })
         xs={6}
         md={3}
       >
-        <FieldMultiSelectionFinal
-          ariaLabelledBy="organizationFormTypesLabel"
-          dataOptions={typeOptions}
-          filter={filter}
-          formatter={formatter}
-          id="organizations-type"
-          itemToString={itemToString}
-          label={<FormattedMessage id="ui-organizations.summary.type" />}
-          name="organizationTypes"
-          onChange={getSelectedTypes}
+        <FieldOrganizationTypes
+          organizationTypes={_organizationTypes}
+          mutators={mutators}
         />
       </Col>
       <Col
